requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
idna==3.3
jwt==1.3.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
requests==2.28.0
urllib3==1.26.9
colorama~=0.4.6
pytest>=6.2.5
pytest-cov>=4.0.0
pytest-subtests>=0.8.0